move simplex functors to their own files
