+++ src/wmtk/simplex/CMakeLists.txt
@@ -4,5 +4,7 @@ set(SRC_FILES
     Simplex.hpp
     SimplexCollection.hpp
     SimplexCollection.cpp
+    SimplexEqualFunctor.hpp
+    SimplexLessFunctor.hpp
 )
 target_sources(wildmeshing_toolkit PRIVATE ${SRC_FILES})
